requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.17.1
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blis==0.7.11
breadability==0.1.20
build==1.2.1
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
coloredlogs==15.0.1
confection==0.1.5
cryptography==43.0.0
cymem==2.0.8
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
docopt==0.6.2
duckduckgo_search==6.2.1
exa-py==1.0.16
faiss-cpu==1.8.0.post1
fake-useragent==1.5.1
fastapi==0.110.3
filelock==3.15.4
flatbuffers==24.3.25
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google_search_results==2.4.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==30.1.0
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.52
langchain-exa==0.0.1
langchain-experimental==0.0.53
langchain-openai==0.0.6
langchain-text-splitters==0.0.1
langchainhub==0.1.20
langcodes==3.4.0
langsmith==0.1.93
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.2.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
multitasking==0.0.11
murmurhash==1.0.10
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.1.9
onnxruntime==1.18.1
openagi==0.2.9.1b2
openai==1.37.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.1
posthog==3.5.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreqwest_impersonate==0.5.2
pytest==8.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacytextblob==4.0.0
SQLAlchemy==2.0.31
srsly==2.4.8
starlette==0.37.2
sumy==0.11.0
sympy==1.13.1
tblib==3.0.0
tenacity==8.5.0
textblob==0.15.3
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.4
typer==0.9.4
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.29.0
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
xorbits==0.7.2
xoscar==0.3.2
yarl==1.9.4
yfinance==0.2.41
zipp==3.19.2
